+++ CHANGES.md
@@ -4,7 +4,8 @@
 2.0b10 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Add `Requirements` component that can create a `requirements.lock` for the
+  `AppEnv` component.
 
 
 2.0b9 (2020-05-09)